# api/openapi.yaml
openapi: "3.0.3"
info:
  title: gin-vagrant-demo API
  version: "1"
  description: |
    Task manager API — a Gin demo evolved into a product.

    **Auth flow:** call `POST /v1/auth/token` with your client credentials to get a JWT,
    then pass it as `Authorization: Bearer <token>` on all `/v1/items` requests.

servers:
  - url: http://localhost:8080
    description: Local / Vagrant VM

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Item:
      type: object
      properties:
        id:          { type: integer, example: 1 }
        title:       { type: string,  example: "Buy milk" }
        body:        { type: string,  example: "2% from the corner store" }
        owner_id:    { type: string,  example: "demo" }
        created_at:  { type: string,  format: date-time }
        updated_at:  { type: string,  format: date-time }

    Error:
      type: object
      properties:
        error: { type: string, example: "not found" }

  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Caller does not own this item
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Item not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }

paths:
  /livez:
    get:
      summary: Liveness probe — is the process alive?
      tags: [infra]
      responses:
        "200": { description: Process alive, content: { application/json: { schema: { type: object, properties: { status: { type: string, example: ok } } } } } }

  /readyz:
    get:
      summary: Readiness probe — is the backing store reachable?
      tags: [infra]
      responses:
        "200": { description: Ready to serve traffic }
        "503": { description: Backing store unreachable }

  /health:
    get:
      summary: Backward-compatible liveness alias (used by Docker HEALTHCHECK and vm.sh)
      tags: [infra]
      responses:
        "200": { description: Ok }

  /version:
    get:
      summary: Build metadata
      tags: [infra]
      responses:
        "200":
          description: Build info stamped at compile time
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:    { type: string }
                  commit:     { type: string }
                  build_time: { type: string, format: date-time }

  /v1/auth/token:
    post:
      summary: Exchange client credentials for a 24-hour JWT
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [client_id, client_secret]
              properties:
                client_id:     { type: string, example: demo }
                client_secret: { type: string, example: demo-secret }
      responses:
        "200":
          description: JWT issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:      { type: string }
                  expires_in: { type: integer, example: 86400 }
        "401": { $ref: '#/components/responses/Unauthorized' }

  /v1/items:
    get:
      summary: List items owned by the authenticated caller
      tags: [items]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: Paginated item list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:  { type: array, items: { $ref: '#/components/schemas/Item' } }
                  limit:  { type: integer }
                  offset: { type: integer }
        "401": { $ref: '#/components/responses/Unauthorized' }

    post:
      summary: Create an item
      tags: [items]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, minLength: 1, maxLength: 255 }
                body:  { type: string, maxLength: 65535 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Item' }
        "400": { description: Validation error }
        "401": { $ref: '#/components/responses/Unauthorized' }

  /v1/items/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }

    get:
      summary: Get an item by ID
      tags: [items]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Item' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }

    put:
      summary: Replace an item
      tags: [items]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, minLength: 1, maxLength: 255 }
                body:  { type: string, maxLength: 65535 }
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Item' }
        "400": { description: Validation error }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }

    delete:
      summary: Delete an item
      tags: [items]
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Deleted }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }